fire shot endpoint better
