+++ game_endpoint.py
@@ -1,4 +1,4 @@
-from utils import ExceptionWithResponseDict
+from utils import ExceptionWithResponseDict, save_game
 
 from battle_ship.game import get_battle_ship_game
 
@@ -13,17 +13,6 @@ class ArgumentException(ExceptionWithResponseDict):
             "status": "error",
             "type": "ArgumentException",
             "message": msg
-        }
-
-        self.resp_dict = err
-
-
-class GameAlreadyExistsException(ExceptionWithResponseDict):
-    def __init__(self, name):
-        err = {
-            "status": "error",
-            "type": "GameAlreadyExistsException",
-            "message": "Game {} already exisits".format(name)
         }
 
         self.resp_dict = err
@@ -65,21 +54,13 @@ class CreateGameEndpoint:
         return game_name, active, inactive
 
     def create_game(self, name, active, inactive):
-        game_file_path = 'games/{}.json'.format(name)
-
-        if not os.path.exists('games'):
-            os.makedirs('games')
-
-        if os.path.exists(game_file_path):
-            raise GameAlreadyExistsException(name)
-
         game_dict = get_battle_ship_game(
             name,
             active,
             inactive
         ).to_dict()
 
-        self.save_to_file(game_file_path, game_dict)
+        save_game(game_dict)
 
         return {
             "status": "success",
@@ -89,7 +70,3 @@ class CreateGameEndpoint:
                 inactive
             )
         }
-
-    def save_to_file(self, game_file_path, game_dict):
-        with open(game_file_path, 'w') as f:
-            f.write(json.dumps(game_dict))
